requirements.txt
click==8.0.4
colorama==0.4.4
joblib==1.1.0
nltk==3.7
numpy==1.22.3
regex==2022.3.2
scikit-learn==1.0.2
scipy==1.8.0
threadpoolctl==3.1.0
tqdm==4.63.0
newspaper3k==0.2.8
